+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+### Unreleased
+
+ * `NftId.[to|from]string()` now uses format `1.2.3/4` instead of `1.2.3@4`
+ * `TokenNftInfoQuery.setNftId()`
+ * `TokenNftInfoQuery.byNftId()` - Use `TokenNftInfoQuery.setNftId()` instead
+ * `TokenNftInfoQuery.byAccountId()` with no replacement
+ * `TokenNftInfoQuery.byTokenId()` with no replacement
+ * `TokenNftInfoQuery.[set|get]Start()` with no replacement
+ * `TokenNftInfoQuery.[set|get]End()` with no replacement
+
 ### v2.0.13
 
 ### Added
@@ -14,13 +24,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * `TokenAssociation` and `TransactionRecord.automaticTokenAssociations`
  * `AccountInfo.maxAutomaticTokenAssociations`
  * `TokenRelationship.automaticAssociation`
+ * `TokenNftInfoQuery.setNftId()`
  * New status codes
+
+### Deprecated
+
+ * `TokenNftInfoQuery.[by|get]AccountId()` with no replacement
+ * `TokenNftInfoQuery.[by|get]TokenId()` with no replacement
+ * `TokenNftInfoQuery.[set|get]Start()` with no replacement
+ * `TokenNftInfoQuery.[set|get]End()` with no replacement
+ * `TokenNftInfoQuery.byNftId()` use `.setNftId()` instead
 
 ### Fixed
 
  * TLS connector failing when the networks address book did not have cert hashes
 
-### v2.0.12
+
+## v2.0.12
 
 ### Added
 
@@ -40,7 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  * Deprecated `Client.setMax[TransactionFee|QueryPayment]()`, added `Client.setDefaultMax[TransactionFee|QueryPayment]()` and `Client.getDefaultMax[TransactionFee|QueryPayment]()`
 
-### v2.0.11
+## v2.0.11
 
 ### Added